code clear
