--- configs/detection/fcos/fcos2.py
@@ -1,59 +1,0 @@
-# model settings
-model = dict(
-    type='FCOS',
-    pretrained=
-    '/home/yunji.cjy/.cache/torch/hub/checkpoints/resnet50_msra-5891d200.pth',
-    backbone=dict(
-        type='ResNet',
-        depth=50,
-        num_stages=4,
-        out_indices=(1, 2, 3, 4),
-        frozen_stages=1,
-        norm_cfg=dict(type='BN', requires_grad=False),
-        norm_eval=True,
-        style='caffe'),
-    neck=dict(
-        type='FPN',
-        in_channels=[256, 512, 1024, 2048],
-        out_channels=256,
-        start_level=1,
-        add_extra_convs='on_output',  # use P5
-        num_outs=5,
-        relu_before_extra_convs=True),
-    head=dict(
-        type='FCOSHead2',
-        num_classes=80,
-        in_channels=256,
-        stacked_convs=4,
-        feat_channels=256,
-        strides=[8, 16, 32, 64, 128],
-        center_sampling=True,
-        center_sample_radius=1.5,
-        norm_on_bbox=True,
-        centerness_on_reg=True,
-        conv_cfg=None,
-        dcn_on_last_conv=False,
-        loss_cls=dict(
-            type='FocalLoss',
-            use_sigmoid=True,
-            gamma=2.0,
-            alpha=0.25,
-            loss_weight=1.0),
-        loss_bbox=dict(type='GIoULoss', loss_weight=1.0),
-        loss_centerness=dict(
-            type='CrossEntropyLoss', use_sigmoid=True, loss_weight=1.0),
-        norm_cfg=dict(type='GN', num_groups=32, requires_grad=True),
-        init_cfg=dict(
-            type='Normal',
-            layer='Conv2d',
-            std=0.01,
-            override=dict(
-                type='Normal', name='conv_cls', std=0.01, bias_prob=0.01)),
-        conv_bias=True,
-        bbox_coder=dict(type='DistancePointBBoxCoder'),
-        test_cfg=dict(
-            nms_pre=1000,
-            min_bbox_size=0,
-            score_thr=0.05,
-            nms=dict(type='nms', iou_threshold=0.6),
-            max_per_img=100)))
