+++ easycv/models/detection/fcos/__init__.py
@@ -1,5 +1,3 @@
 from .fcos import FCOS
 from .fcos_head import FCOSHead
-from .fcos_head2 import FCOSHead2
-from .fcos_outputs import FCOSOutputs
 from .fpn import FPN
