--- easycv/models/detection/fcos/fcos_head2.py
@@ -1,880 +1,0 @@
-# Copyright (c) Facebook, Inc. and its affiliates. All Rights Reserved
-import math
-import warnings
-
-import numpy as np
-import torch
-import torch.nn as nn
-import torch.nn.functional as F
-from mmcv.cnn import ConvModule, Scale
-from mmcv.ops import batched_nms
-from mmdet.core import bbox2result, build_bbox_coder, multi_apply, reduce_mean
-from mmdet.core.anchor.point_generator import MlvlPointGenerator
-from mmdet.core.utils import filter_scores_and_topk, select_single_mlvl
-
-from easycv.models.builder import HEADS, build_loss
-
-INF = 1e8
-
-
-@HEADS.register_module()
-class FCOSHead2(nn.Module):
-    """Implements the DETR transformer head.
-    See `paper: End-to-End Object Detection with Transformers
-    <https://arxiv.org/pdf/2005.12872>`_ for details.
-    Args:
-        num_classes (int): Number of categories excluding the background.
-    """
-
-    def __init__(self,
-                 num_classes,
-                 in_channels,
-                 stacked_convs=4,
-                 feat_channels=256,
-                 strides=[8, 16, 32, 64, 128],
-                 regress_ranges=((-1, 64), (64, 128), (128, 256), (256, 512),
-                                 (512, INF)),
-                 center_sampling=False,
-                 center_sample_radius=1.5,
-                 norm_on_bbox=False,
-                 centerness_on_reg=False,
-                 conv_cfg=None,
-                 dcn_on_last_conv=False,
-                 loss_cls=dict(
-                     type='FocalLoss',
-                     use_sigmoid=True,
-                     gamma=2.0,
-                     alpha=0.25,
-                     loss_weight=1.0),
-                 loss_bbox=dict(type='IoULoss', loss_weight=1.0),
-                 loss_centerness=dict(
-                     type='CrossEntropyLoss',
-                     use_sigmoid=True,
-                     loss_weight=1.0),
-                 norm_cfg=dict(type='GN', num_groups=32, requires_grad=True),
-                 init_cfg=dict(
-                     type='Normal',
-                     layer='Conv2d',
-                     std=0.01,
-                     override=dict(
-                         type='Normal',
-                         name='conv_cls',
-                         std=0.01,
-                         bias_prob=0.01)),
-                 conv_bias=True,
-                 bbox_coder=dict(type='DistancePointBBoxCoder'),
-                 test_cfg=dict(
-                     nms_pre=1000,
-                     min_bbox_size=0,
-                     score_thr=0.05,
-                     nms=dict(type='nms', iou_threshold=0.5),
-                     max_per_img=100),
-                 **kwargs):
-        super(FCOSHead2, self).__init__()
-        self.regress_ranges = regress_ranges
-        self.center_sampling = center_sampling
-        self.center_sample_radius = center_sample_radius
-        self.norm_on_bbox = norm_on_bbox
-        self.centerness_on_reg = centerness_on_reg
-
-        self.num_classes = num_classes
-        self.use_sigmoid_cls = loss_cls.get('use_sigmoid', False)
-        if self.use_sigmoid_cls:
-            self.cls_out_channels = num_classes
-        else:
-            self.cls_out_channels = num_classes + 1
-        self.in_channels = in_channels
-        self.feat_channels = feat_channels
-        self.stacked_convs = stacked_convs
-        self.strides = strides
-        self.dcn_on_last_conv = dcn_on_last_conv
-        assert conv_bias == 'auto' or isinstance(conv_bias, bool)
-        self.conv_bias = conv_bias
-        self.loss_cls = build_loss(loss_cls)
-        self.loss_bbox = build_loss(loss_bbox)
-        self.bbox_coder = build_bbox_coder(bbox_coder)
-
-        self.prior_generator = MlvlPointGenerator(strides)
-
-        # In order to keep a more general interface and be consistent with
-        # anchor_head. We can think of point like one anchor
-        self.num_base_priors = self.prior_generator.num_base_priors[0]
-
-        self.test_cfg = test_cfg
-        self.conv_cfg = conv_cfg
-        self.norm_cfg = norm_cfg
-        self.fp16_enabled = False
-
-        self._init_layers()
-
-        self.loss_centerness = build_loss(loss_centerness)
-
-    def _init_layers(self):
-        """Initialize layers of the head."""
-        self._init_cls_convs()
-        self._init_reg_convs()
-        self._init_predictor()
-        self.conv_centerness = nn.Conv2d(self.feat_channels, 1, 3, padding=1)
-        self.scales = nn.ModuleList([Scale(1.0) for _ in self.strides])
-
-    def _init_cls_convs(self):
-        """Initialize classification conv layers of the head."""
-        self.cls_convs = nn.ModuleList()
-        for i in range(self.stacked_convs):
-            chn = self.in_channels if i == 0 else self.feat_channels
-            if self.dcn_on_last_conv and i == self.stacked_convs - 1:
-                conv_cfg = dict(type='DCNv2')
-            else:
-                conv_cfg = self.conv_cfg
-            self.cls_convs.append(
-                ConvModule(
-                    chn,
-                    self.feat_channels,
-                    3,
-                    stride=1,
-                    padding=1,
-                    conv_cfg=conv_cfg,
-                    norm_cfg=self.norm_cfg,
-                    bias=self.conv_bias))
-
-    def _init_reg_convs(self):
-        """Initialize bbox regression conv layers of the head."""
-        self.reg_convs = nn.ModuleList()
-        for i in range(self.stacked_convs):
-            chn = self.in_channels if i == 0 else self.feat_channels
-            if self.dcn_on_last_conv and i == self.stacked_convs - 1:
-                conv_cfg = dict(type='DCNv2')
-            else:
-                conv_cfg = self.conv_cfg
-            self.reg_convs.append(
-                ConvModule(
-                    chn,
-                    self.feat_channels,
-                    3,
-                    stride=1,
-                    padding=1,
-                    conv_cfg=conv_cfg,
-                    norm_cfg=self.norm_cfg,
-                    bias=self.conv_bias))
-
-    def _init_predictor(self):
-        """Initialize predictor layers of the head."""
-        self.conv_cls = nn.Conv2d(
-            self.feat_channels, self.cls_out_channels, 3, padding=1)
-        self.conv_reg = nn.Conv2d(self.feat_channels, 4, 3, padding=1)
-
-    def init_weights(self):
-        for m in self.modules():
-            if isinstance(m, nn.Conv2d):
-                torch.nn.init.normal_(m.weight, std=0.01)
-                if hasattr(m, 'bias') and m.bias is not None:
-                    nn.init.constant_(m.bias, 0)
-
-        # initialize the bias for focal loss
-        prior_prob = 0.01
-        bias_value = -math.log((1 - prior_prob) / prior_prob)
-        torch.nn.init.constant_(self.conv_cls.bias, bias_value)
-
-    def forward(self, feats):
-        """Forward features from the upstream network.
-
-        Args:
-            feats (tuple[Tensor]): Features from the upstream network, each is
-                a 4D-tensor.
-
-        Returns:
-            tuple:
-                cls_scores (list[Tensor]): Box scores for each scale level, \
-                    each is a 4D-tensor, the channel number is \
-                    num_points * num_classes.
-                bbox_preds (list[Tensor]): Box energies / deltas for each \
-                    scale level, each is a 4D-tensor, the channel number is \
-                    num_points * 4.
-                centernesses (list[Tensor]): centerness for each scale level, \
-                    each is a 4D-tensor, the channel number is num_points * 1.
-        """
-        return multi_apply(self.forward_single, feats, self.scales,
-                           self.strides)
-
-    def forward_single(self, x, scale, stride):
-        """Forward features of a single scale level.
-
-        Args:
-            x (Tensor): FPN feature maps of the specified stride.
-            scale (:obj: `mmcv.cnn.Scale`): Learnable scale module to resize
-                the bbox prediction.
-            stride (int): The corresponding stride for feature maps, only
-                used to normalize the bbox prediction when self.norm_on_bbox
-                is True.
-
-        Returns:
-            tuple: scores for each class, bbox predictions and centerness \
-                predictions of input feature maps.
-        """
-        cls_feat = x
-        reg_feat = x
-
-        for cls_layer in self.cls_convs:
-            cls_feat = cls_layer(cls_feat)
-        cls_score = self.conv_cls(cls_feat)
-
-        for reg_layer in self.reg_convs:
-            reg_feat = reg_layer(reg_feat)
-        bbox_pred = self.conv_reg(reg_feat)
-
-        if self.centerness_on_reg:
-            centerness = self.conv_centerness(reg_feat)
-        else:
-            centerness = self.conv_centerness(cls_feat)
-        # scale the bbox_pred of different level
-        # float to avoid overflow when enabling FP16
-        bbox_pred = scale(bbox_pred).float()
-        if self.norm_on_bbox:
-            # bbox_pred needed for gradient computation has been modified
-            # by F.relu(bbox_pred) when run with PyTorch 1.10. So replace
-            # F.relu(bbox_pred) with bbox_pred.clamp(min=0)
-            bbox_pred = bbox_pred.clamp(min=0)
-            if not self.training:
-                bbox_pred *= stride
-        else:
-            bbox_pred = bbox_pred.exp()
-        return cls_score, bbox_pred, centerness
-
-    def forward_train(self,
-                      x,
-                      img_metas,
-                      gt_bboxes,
-                      gt_labels=None,
-                      gt_bboxes_ignore=None,
-                      proposal_cfg=None,
-                      **kwargs):
-        outs = self.forward(x)
-        if gt_labels is None:
-            loss_inputs = outs + (gt_bboxes, img_metas)
-        else:
-            loss_inputs = outs + (gt_bboxes, gt_labels, img_metas)
-        losses = self.loss(*loss_inputs, gt_bboxes_ignore=gt_bboxes_ignore)
-        return losses
-
-    def forward_test(self, feats, img_metas, rescale=False):
-        """Test function without test-time augmentation.
-
-        Args:
-            feats (tuple[torch.Tensor]): Multi-level features from the
-                upstream network, each is a 4D-tensor.
-            img_metas (list[dict]): List of image information.
-            rescale (bool, optional): Whether to rescale the results.
-                Defaults to False.
-
-        Returns:
-            list[tuple[Tensor, Tensor]]: Each item in result_list is 2-tuple.
-                The first item is ``bboxes`` with shape (n, 5),
-                where 5 represent (tl_x, tl_y, br_x, br_y, score).
-                The shape of the second tensor in the tuple is ``labels``
-                with shape (n, ).
-        """
-        # print(feats, feats.shape)
-        outs = self.forward(feats)
-        # print(outs)
-        results_list = self.get_bboxes(
-            *outs, img_metas=img_metas, rescale=True)
-        # print(results)
-        results = [
-            bbox2result(det_bboxes, det_labels, self.num_classes)
-            for det_bboxes, det_labels in results_list
-        ]
-        # print(results)
-
-        detection_boxes = []
-        detection_scores = []
-        detection_classes = []
-        for res_i in results:
-            bbox_result = res_i
-            bboxes = np.vstack(bbox_result)
-            labels = [
-                np.full(bbox.shape[0], i, dtype=np.int32)
-                for i, bbox in enumerate(bbox_result)
-            ]
-            labels = np.concatenate(labels)
-
-            scores = bboxes[:, 4] if bboxes.shape[1] == 5 else None
-            bboxes = bboxes[:, 0:4] if bboxes.shape[1] == 5 else bboxes
-            assert bboxes.shape[1] == 4
-
-            detection_boxes.append(bboxes)
-            detection_scores.append(scores)
-            detection_classes.append(labels)
-
-        assert len(img_metas) == 1
-        outputs = {
-            'detection_boxes': detection_boxes,
-            'detection_scores': detection_scores,
-            'detection_classes': detection_classes,
-            'img_metas': img_metas
-        }
-
-        return outputs
-
-    def loss(self,
-             cls_scores,
-             bbox_preds,
-             centernesses,
-             gt_bboxes,
-             gt_labels,
-             img_metas,
-             gt_bboxes_ignore=None):
-        """Compute loss of the head.
-
-        Args:
-            cls_scores (list[Tensor]): Box scores for each scale level,
-                each is a 4D-tensor, the channel number is
-                num_points * num_classes.
-            bbox_preds (list[Tensor]): Box energies / deltas for each scale
-                level, each is a 4D-tensor, the channel number is
-                num_points * 4.
-            centernesses (list[Tensor]): centerness for each scale level, each
-                is a 4D-tensor, the channel number is num_points * 1.
-            gt_bboxes (list[Tensor]): Ground truth bboxes for each image with
-                shape (num_gts, 4) in [tl_x, tl_y, br_x, br_y] format.
-            gt_labels (list[Tensor]): class indices corresponding to each box
-            img_metas (list[dict]): Meta information of each image, e.g.,
-                image size, scaling factor, etc.
-            gt_bboxes_ignore (None | list[Tensor]): specify which bounding
-                boxes can be ignored when computing the loss.
-
-        Returns:
-            dict[str, Tensor]: A dictionary of loss components.
-        """
-        assert len(cls_scores) == len(bbox_preds) == len(centernesses)
-        featmap_sizes = [featmap.size()[-2:] for featmap in cls_scores]
-        all_level_points = self.prior_generator.grid_priors(
-            featmap_sizes,
-            dtype=bbox_preds[0].dtype,
-            device=bbox_preds[0].device)
-        labels, bbox_targets = self.get_targets(all_level_points, gt_bboxes,
-                                                gt_labels)
-        # print(labels, bbox_targets)
-        # exit()
-
-        num_imgs = cls_scores[0].size(0)
-        # flatten cls_scores, bbox_preds and centerness
-        flatten_cls_scores = [
-            cls_score.permute(0, 2, 3, 1).reshape(-1, self.cls_out_channels)
-            for cls_score in cls_scores
-        ]
-        flatten_bbox_preds = [
-            bbox_pred.permute(0, 2, 3, 1).reshape(-1, 4)
-            for bbox_pred in bbox_preds
-        ]
-        flatten_centerness = [
-            centerness.permute(0, 2, 3, 1).reshape(-1)
-            for centerness in centernesses
-        ]
-        flatten_cls_scores = torch.cat(flatten_cls_scores)
-        flatten_bbox_preds = torch.cat(flatten_bbox_preds)
-        flatten_centerness = torch.cat(flatten_centerness)
-        flatten_labels = torch.cat(labels)
-        flatten_bbox_targets = torch.cat(bbox_targets)
-        # repeat points to align with bbox_preds
-        flatten_points = torch.cat(
-            [points.repeat(num_imgs, 1) for points in all_level_points])
-
-        # FG cat_id: [0, num_classes -1], BG cat_id: num_classes
-        bg_class_ind = self.num_classes
-        pos_inds = ((flatten_labels >= 0)
-                    & (flatten_labels < bg_class_ind)).nonzero().reshape(-1)
-        num_pos = torch.tensor(
-            len(pos_inds), dtype=torch.float, device=bbox_preds[0].device)
-        num_pos = max(reduce_mean(num_pos), 1.0)
-        loss_cls = self.loss_cls(
-            flatten_cls_scores, flatten_labels, avg_factor=num_pos)
-
-        pos_bbox_preds = flatten_bbox_preds[pos_inds]
-        pos_centerness = flatten_centerness[pos_inds]
-        pos_bbox_targets = flatten_bbox_targets[pos_inds]
-        pos_centerness_targets = self.centerness_target(pos_bbox_targets)
-        # print(pos_centerness_targets)
-        # exit()
-        # centerness weighted iou loss
-        centerness_denorm = max(
-            reduce_mean(pos_centerness_targets.sum().detach()), 1e-6)
-
-        if len(pos_inds) > 0:
-            pos_points = flatten_points[pos_inds]
-            pos_decoded_bbox_preds = self.bbox_coder.decode(
-                pos_points, pos_bbox_preds)
-            pos_decoded_target_preds = self.bbox_coder.decode(
-                pos_points, pos_bbox_targets)
-            loss_bbox = self.loss_bbox(
-                pos_decoded_bbox_preds,
-                pos_decoded_target_preds,
-                weight=pos_centerness_targets,
-                avg_factor=centerness_denorm)
-            loss_centerness = self.loss_centerness(
-                pos_centerness, pos_centerness_targets, avg_factor=num_pos)
-        else:
-            loss_bbox = pos_bbox_preds.sum()
-            loss_centerness = pos_centerness.sum()
-
-        return dict(
-            loss_cls=loss_cls,
-            loss_bbox=loss_bbox,
-            loss_centerness=loss_centerness)
-
-    def get_targets(self, points, gt_bboxes_list, gt_labels_list):
-        """Compute regression, classification and centerness targets for points
-        in multiple images.
-
-        Args:
-            points (list[Tensor]): Points of each fpn level, each has shape
-                (num_points, 2).
-            gt_bboxes_list (list[Tensor]): Ground truth bboxes of each image,
-                each has shape (num_gt, 4).
-            gt_labels_list (list[Tensor]): Ground truth labels of each box,
-                each has shape (num_gt,).
-
-        Returns:
-            tuple:
-                concat_lvl_labels (list[Tensor]): Labels of each level. \
-                concat_lvl_bbox_targets (list[Tensor]): BBox targets of each \
-                    level.
-        """
-        assert len(points) == len(self.regress_ranges)
-        num_levels = len(points)
-        # expand regress ranges to align with points
-        expanded_regress_ranges = [
-            points[i].new_tensor(self.regress_ranges[i])[None].expand_as(
-                points[i]) for i in range(num_levels)
-        ]
-        # concat all levels points and regress ranges
-        concat_regress_ranges = torch.cat(expanded_regress_ranges, dim=0)
-        concat_points = torch.cat(points, dim=0)
-
-        # the number of points per img, per lvl
-        num_points = [center.size(0) for center in points]
-
-        # get labels and bbox_targets of each image
-        labels_list, bbox_targets_list = multi_apply(
-            self._get_target_single,
-            gt_bboxes_list,
-            gt_labels_list,
-            points=concat_points,
-            regress_ranges=concat_regress_ranges,
-            num_points_per_lvl=num_points)
-
-        # split to per img, per level
-        labels_list = [labels.split(num_points, 0) for labels in labels_list]
-        bbox_targets_list = [
-            bbox_targets.split(num_points, 0)
-            for bbox_targets in bbox_targets_list
-        ]
-
-        # concat per level image
-        concat_lvl_labels = []
-        concat_lvl_bbox_targets = []
-        for i in range(num_levels):
-            concat_lvl_labels.append(
-                torch.cat([labels[i] for labels in labels_list]))
-            bbox_targets = torch.cat(
-                [bbox_targets[i] for bbox_targets in bbox_targets_list])
-            if self.norm_on_bbox:
-                bbox_targets = bbox_targets / self.strides[i]
-            concat_lvl_bbox_targets.append(bbox_targets)
-        return concat_lvl_labels, concat_lvl_bbox_targets
-
-    def _get_target_single(self, gt_bboxes, gt_labels, points, regress_ranges,
-                           num_points_per_lvl):
-        """Compute regression and classification targets for a single image."""
-        num_points = points.size(0)
-        num_gts = gt_labels.size(0)
-        if num_gts == 0:
-            return gt_labels.new_full((num_points,), self.num_classes), \
-                   gt_bboxes.new_zeros((num_points, 4))
-
-        areas = (gt_bboxes[:, 2] - gt_bboxes[:, 0]) * (
-            gt_bboxes[:, 3] - gt_bboxes[:, 1])
-        # TODO: figure out why these two are different
-        # areas = areas[None].expand(num_points, num_gts)
-        areas = areas[None].repeat(num_points, 1)
-        regress_ranges = regress_ranges[:, None, :].expand(
-            num_points, num_gts, 2)
-        gt_bboxes = gt_bboxes[None].expand(num_points, num_gts, 4)
-        xs, ys = points[:, 0], points[:, 1]
-        xs = xs[:, None].expand(num_points, num_gts)
-        ys = ys[:, None].expand(num_points, num_gts)
-
-        left = xs - gt_bboxes[..., 0]
-        right = gt_bboxes[..., 2] - xs
-        top = ys - gt_bboxes[..., 1]
-        bottom = gt_bboxes[..., 3] - ys
-        bbox_targets = torch.stack((left, top, right, bottom), -1)
-
-        if self.center_sampling:
-            # condition1: inside a `center bbox`
-            radius = self.center_sample_radius
-            center_xs = (gt_bboxes[..., 0] + gt_bboxes[..., 2]) / 2
-            center_ys = (gt_bboxes[..., 1] + gt_bboxes[..., 3]) / 2
-            center_gts = torch.zeros_like(gt_bboxes)
-            stride = center_xs.new_zeros(center_xs.shape)
-
-            # project the points on current lvl back to the `original` sizes
-            lvl_begin = 0
-            for lvl_idx, num_points_lvl in enumerate(num_points_per_lvl):
-                lvl_end = lvl_begin + num_points_lvl
-                stride[lvl_begin:lvl_end] = self.strides[lvl_idx] * radius
-                lvl_begin = lvl_end
-
-            x_mins = center_xs - stride
-            y_mins = center_ys - stride
-            x_maxs = center_xs + stride
-            y_maxs = center_ys + stride
-            center_gts[..., 0] = torch.where(x_mins > gt_bboxes[..., 0],
-                                             x_mins, gt_bboxes[..., 0])
-            center_gts[..., 1] = torch.where(y_mins > gt_bboxes[..., 1],
-                                             y_mins, gt_bboxes[..., 1])
-            center_gts[..., 2] = torch.where(x_maxs > gt_bboxes[..., 2],
-                                             gt_bboxes[..., 2], x_maxs)
-            center_gts[..., 3] = torch.where(y_maxs > gt_bboxes[..., 3],
-                                             gt_bboxes[..., 3], y_maxs)
-
-            cb_dist_left = xs - center_gts[..., 0]
-            cb_dist_right = center_gts[..., 2] - xs
-            cb_dist_top = ys - center_gts[..., 1]
-            cb_dist_bottom = center_gts[..., 3] - ys
-            center_bbox = torch.stack(
-                (cb_dist_left, cb_dist_top, cb_dist_right, cb_dist_bottom), -1)
-            inside_gt_bbox_mask = center_bbox.min(-1)[0] > 0
-        else:
-            # condition1: inside a gt bbox
-            inside_gt_bbox_mask = bbox_targets.min(-1)[0] > 0
-
-        # condition2: limit the regression range for each location
-        max_regress_distance = bbox_targets.max(-1)[0]
-        inside_regress_range = (
-            (max_regress_distance >= regress_ranges[..., 0])
-            & (max_regress_distance <= regress_ranges[..., 1]))
-
-        # if there are still more than one objects for a location,
-        # we choose the one with minimal area
-        areas[inside_gt_bbox_mask == 0] = INF
-        areas[inside_regress_range == 0] = INF
-        min_area, min_area_inds = areas.min(dim=1)
-
-        labels = gt_labels[min_area_inds]
-        labels[min_area == INF] = self.num_classes  # set as BG
-        bbox_targets = bbox_targets[range(num_points), min_area_inds]
-
-        return labels, bbox_targets
-
-    def centerness_target(self, pos_bbox_targets):
-        """Compute centerness targets.
-
-        Args:
-            pos_bbox_targets (Tensor): BBox targets of positive bboxes in shape
-                (num_pos, 4)
-
-        Returns:
-            Tensor: Centerness target.
-        """
-        # only calculate pos centerness targets, otherwise there may be nan
-        left_right = pos_bbox_targets[:, [0, 2]]
-        top_bottom = pos_bbox_targets[:, [1, 3]]
-        if len(left_right) == 0:
-            centerness_targets = left_right[..., 0]
-        else:
-            centerness_targets = (
-                left_right.min(dim=-1)[0] / left_right.max(dim=-1)[0]) * (
-                    top_bottom.min(dim=-1)[0] / top_bottom.max(dim=-1)[0])
-        return torch.sqrt(centerness_targets)
-
-    def _get_points_single(self,
-                           featmap_size,
-                           stride,
-                           dtype,
-                           device,
-                           flatten=False):
-        """Get points according to feature map size.
-
-        This function will be deprecated soon.
-        """
-        warnings.warn(
-            '`_get_points_single` in `FCOSHead` will be '
-            'deprecated soon, we support a multi level point generator now'
-            'you can get points of a single level feature map '
-            'with `self.prior_generator.single_level_grid_priors` ')
-
-        y, x = super()._get_points_single(featmap_size, stride, dtype, device)
-        points = torch.stack((x.reshape(-1) * stride, y.reshape(-1) * stride),
-                             dim=-1) + stride // 2
-        return points
-
-    def get_bboxes(self,
-                   cls_scores,
-                   bbox_preds,
-                   score_factors=None,
-                   img_metas=None,
-                   cfg=None,
-                   rescale=False,
-                   with_nms=True,
-                   **kwargs):
-        """Transform network outputs of a batch into bbox results.
-
-        Note: When score_factors is not None, the cls_scores are
-        usually multiplied by it then obtain the real score used in NMS,
-        such as CenterNess in FCOS, IoU branch in ATSS.
-
-        Args:
-            cls_scores (list[Tensor]): Classification scores for all
-                scale levels, each is a 4D-tensor, has shape
-                (batch_size, num_priors * num_classes, H, W).
-            bbox_preds (list[Tensor]): Box energies / deltas for all
-                scale levels, each is a 4D-tensor, has shape
-                (batch_size, num_priors * 4, H, W).
-            score_factors (list[Tensor], Optional): Score factor for
-                all scale level, each is a 4D-tensor, has shape
-                (batch_size, num_priors * 1, H, W). Default None.
-            img_metas (list[dict], Optional): Image meta info. Default None.
-            cfg (mmcv.Config, Optional): Test / postprocessing configuration,
-                if None, test_cfg would be used.  Default None.
-            rescale (bool): If True, return boxes in original image space.
-                Default False.
-            with_nms (bool): If True, do nms before return boxes.
-                Default True.
-
-        Returns:
-            list[list[Tensor, Tensor]]: Each item in result_list is 2-tuple.
-                The first item is an (n, 5) tensor, where the first 4 columns
-                are bounding box positions (tl_x, tl_y, br_x, br_y) and the
-                5-th column is a score between 0 and 1. The second item is a
-                (n,) tensor where each item is the predicted class label of
-                the corresponding box.
-        """
-        assert len(cls_scores) == len(bbox_preds)
-
-        if score_factors is None:
-            # e.g. Retina, FreeAnchor, Foveabox, etc.
-            with_score_factors = False
-        else:
-            # e.g. FCOS, PAA, ATSS, AutoAssign, etc.
-            with_score_factors = True
-            assert len(cls_scores) == len(score_factors)
-
-        num_levels = len(cls_scores)
-
-        featmap_sizes = [cls_scores[i].shape[-2:] for i in range(num_levels)]
-        mlvl_priors = self.prior_generator.grid_priors(
-            featmap_sizes,
-            dtype=cls_scores[0].dtype,
-            device=cls_scores[0].device)
-
-        result_list = []
-
-        for img_id in range(len(img_metas)):
-            img_meta = img_metas[img_id]
-            cls_score_list = select_single_mlvl(cls_scores, img_id)
-            bbox_pred_list = select_single_mlvl(bbox_preds, img_id)
-            if with_score_factors:
-                score_factor_list = select_single_mlvl(score_factors, img_id)
-            else:
-                score_factor_list = [None for _ in range(num_levels)]
-
-            results = self._get_bboxes_single(cls_score_list, bbox_pred_list,
-                                              score_factor_list, mlvl_priors,
-                                              img_meta, cfg, rescale, with_nms,
-                                              **kwargs)
-            result_list.append(results)
-        return result_list
-
-    def _get_bboxes_single(self,
-                           cls_score_list,
-                           bbox_pred_list,
-                           score_factor_list,
-                           mlvl_priors,
-                           img_meta,
-                           cfg,
-                           rescale=False,
-                           with_nms=True,
-                           **kwargs):
-        """Transform outputs of a single image into bbox predictions.
-
-        Args:
-            cls_score_list (list[Tensor]): Box scores from all scale
-                levels of a single image, each item has shape
-                (num_priors * num_classes, H, W).
-            bbox_pred_list (list[Tensor]): Box energies / deltas from
-                all scale levels of a single image, each item has shape
-                (num_priors * 4, H, W).
-            score_factor_list (list[Tensor]): Score factor from all scale
-                levels of a single image, each item has shape
-                (num_priors * 1, H, W).
-            mlvl_priors (list[Tensor]): Each element in the list is
-                the priors of a single level in feature pyramid. In all
-                anchor-based methods, it has shape (num_priors, 4). In
-                all anchor-free methods, it has shape (num_priors, 2)
-                when `with_stride=True`, otherwise it still has shape
-                (num_priors, 4).
-            img_meta (dict): Image meta info.
-            cfg (mmcv.Config): Test / postprocessing configuration,
-                if None, test_cfg would be used.
-            rescale (bool): If True, return boxes in original image space.
-                Default: False.
-            with_nms (bool): If True, do nms before return boxes.
-                Default: True.
-
-        Returns:
-            tuple[Tensor]: Results of detected bboxes and labels. If with_nms
-                is False and mlvl_score_factor is None, return mlvl_bboxes and
-                mlvl_scores, else return mlvl_bboxes, mlvl_scores and
-                mlvl_score_factor. Usually with_nms is False is used for aug
-                test. If with_nms is True, then return the following format
-
-                - det_bboxes (Tensor): Predicted bboxes with shape \
-                    [num_bboxes, 5], where the first 4 columns are bounding \
-                    box positions (tl_x, tl_y, br_x, br_y) and the 5-th \
-                    column are scores between 0 and 1.
-                - det_labels (Tensor): Predicted labels of the corresponding \
-                    box with shape [num_bboxes].
-        """
-        if score_factor_list[0] is None:
-            # e.g. Retina, FreeAnchor, etc.
-            with_score_factors = False
-        else:
-            # e.g. FCOS, PAA, ATSS, etc.
-            with_score_factors = True
-
-        cfg = self.test_cfg if cfg is None else cfg
-        img_shape = img_meta['img_shape']
-        nms_pre = cfg.get('nms_pre', -1)
-
-        mlvl_bboxes = []
-        mlvl_scores = []
-        mlvl_labels = []
-        if with_score_factors:
-            mlvl_score_factors = []
-        else:
-            mlvl_score_factors = None
-        for level_idx, (cls_score, bbox_pred, score_factor, priors) in \
-                enumerate(zip(cls_score_list, bbox_pred_list,
-                              score_factor_list, mlvl_priors)):
-
-            assert cls_score.size()[-2:] == bbox_pred.size()[-2:]
-
-            bbox_pred = bbox_pred.permute(1, 2, 0).reshape(-1, 4)
-            if with_score_factors:
-                score_factor = score_factor.permute(1, 2,
-                                                    0).reshape(-1).sigmoid()
-            cls_score = cls_score.permute(1, 2,
-                                          0).reshape(-1, self.cls_out_channels)
-            if self.use_sigmoid_cls:
-                scores = cls_score.sigmoid()
-            else:
-                # remind that we set FG labels to [0, num_class-1]
-                # since mmdet v2.0
-                # BG cat_id: num_class
-                scores = cls_score.softmax(-1)[:, :-1]
-
-            # After https://github.com/open-mmlab/mmdetection/pull/6268/,
-            # this operation keeps fewer bboxes under the same `nms_pre`.
-            # There is no difference in performance for most models. If you
-            # find a slight drop in performance, you can set a larger
-            # `nms_pre` than before.
-            results = filter_scores_and_topk(
-                scores, cfg.score_thr, nms_pre,
-                dict(bbox_pred=bbox_pred, priors=priors))
-            scores, labels, keep_idxs, filtered_results = results
-
-            bbox_pred = filtered_results['bbox_pred']
-            priors = filtered_results['priors']
-
-            if with_score_factors:
-                score_factor = score_factor[keep_idxs]
-
-            bboxes = self.bbox_coder.decode(
-                priors, bbox_pred, max_shape=img_shape)
-
-            mlvl_bboxes.append(bboxes)
-            mlvl_scores.append(scores)
-            mlvl_labels.append(labels)
-            if with_score_factors:
-                mlvl_score_factors.append(score_factor)
-
-        return self._bbox_post_process(mlvl_scores, mlvl_labels, mlvl_bboxes,
-                                       img_meta['scale_factor'], cfg, rescale,
-                                       with_nms, mlvl_score_factors, **kwargs)
-
-    def _bbox_post_process(self,
-                           mlvl_scores,
-                           mlvl_labels,
-                           mlvl_bboxes,
-                           scale_factor,
-                           cfg,
-                           rescale=False,
-                           with_nms=True,
-                           mlvl_score_factors=None,
-                           **kwargs):
-        """bbox post-processing method.
-
-        The boxes would be rescaled to the original image scale and do
-        the nms operation. Usually `with_nms` is False is used for aug test.
-
-        Args:
-            mlvl_scores (list[Tensor]): Box scores from all scale
-                levels of a single image, each item has shape
-                (num_bboxes, ).
-            mlvl_labels (list[Tensor]): Box class labels from all scale
-                levels of a single image, each item has shape
-                (num_bboxes, ).
-            mlvl_bboxes (list[Tensor]): Decoded bboxes from all scale
-                levels of a single image, each item has shape (num_bboxes, 4).
-            scale_factor (ndarray, optional): Scale factor of the image arange
-                as (w_scale, h_scale, w_scale, h_scale).
-            cfg (mmcv.Config): Test / postprocessing configuration,
-                if None, test_cfg would be used.
-            rescale (bool): If True, return boxes in original image space.
-                Default: False.
-            with_nms (bool): If True, do nms before return boxes.
-                Default: True.
-            mlvl_score_factors (list[Tensor], optional): Score factor from
-                all scale levels of a single image, each item has shape
-                (num_bboxes, ). Default: None.
-
-        Returns:
-            tuple[Tensor]: Results of detected bboxes and labels. If with_nms
-                is False and mlvl_score_factor is None, return mlvl_bboxes and
-                mlvl_scores, else return mlvl_bboxes, mlvl_scores and
-                mlvl_score_factor. Usually with_nms is False is used for aug
-                test. If with_nms is True, then return the following format
-
-                - det_bboxes (Tensor): Predicted bboxes with shape \
-                    [num_bboxes, 5], where the first 4 columns are bounding \
-                    box positions (tl_x, tl_y, br_x, br_y) and the 5-th \
-                    column are scores between 0 and 1.
-                - det_labels (Tensor): Predicted labels of the corresponding \
-                    box with shape [num_bboxes].
-        """
-        assert len(mlvl_scores) == len(mlvl_bboxes) == len(mlvl_labels)
-
-        mlvl_bboxes = torch.cat(mlvl_bboxes)
-        if rescale:
-            mlvl_bboxes /= mlvl_bboxes.new_tensor(scale_factor)
-        mlvl_scores = torch.cat(mlvl_scores)
-        mlvl_labels = torch.cat(mlvl_labels)
-
-        if mlvl_score_factors is not None:
-            # TODO： Add sqrt operation in order to be consistent with
-            #  the paper.
-            mlvl_score_factors = torch.cat(mlvl_score_factors)
-            mlvl_scores = mlvl_scores * mlvl_score_factors
-
-        if with_nms:
-            if mlvl_bboxes.numel() == 0:
-                det_bboxes = torch.cat([mlvl_bboxes, mlvl_scores[:, None]], -1)
-                return det_bboxes, mlvl_labels
-
-            det_bboxes, keep_idxs = batched_nms(mlvl_bboxes, mlvl_scores,
-                                                mlvl_labels, cfg.nms)
-            det_bboxes = det_bboxes[:cfg.max_per_img]
-            det_labels = mlvl_labels[keep_idxs][:cfg.max_per_img]
-            return det_bboxes, det_labels
-        else:
-            return mlvl_bboxes, mlvl_scores, mlvl_labels
